Refactoring Region.jsx: Moved functions out, made Coloring into functions
